refactor: improve structure, error handling, otimization  and progress monitoring

+++ build.py
@@ -5,18 +5,23 @@ import re
 import time
 import psutil
 import json
+import logging
+import threading
 
 # Configurações do Telegram
-CHAT_ID = "SEU_CHAT_ID"
-BOT_TOKEN = "SEU_BOT_TOKEN"
+CHAT_ID = "0000000000"
+BOT_TOKEN = "0000000000:0000000000aLPb6Yn2xCmCLS5EBh_qGBsM"
 TELEGRAM_API_URL = f"https://api.telegram.org/bot{BOT_TOKEN}"
 
 # Configurações de build
-LUNCH_TARGET = "aosp_spes-user"
+LUNCH_TARGET = "aosp_device-ap4a-user"
 BUILD_TARGET = "bacon"
 
 # Configurações do Pixeldrain
-API_KEY = "12345678910"
+BASE_URL = "https://pixeldrain.com/api"
+FILE_URL = "https://pixeldrain.com"
+API_KEY = "00000000-00000-0000-0000-0000000000"
+
 
 # Funções para envio de mensagens ao Telegram
 def send_telegram_message(message):
@@ -26,7 +31,7 @@ def send_telegram_message(message):
         response = requests.post(url, data=data)
         return response.json().get("result", {}).get("message_id")
     except Exception as e:
-        print(f"Erro ao enviar mensagem ao Telegram: {e}")
+        logging.error(f"Erro ao enviar mensagem ao Telegram: {e}")
 
 def edit_telegram_message(message_id, new_text):
     try:
@@ -39,7 +44,7 @@ def edit_telegram_message(message_id, new_text):
         }
         requests.post(url, data=data)
     except Exception as e:
-        print(f"Erro ao editar mensagem no Telegram: {e}")
+        logging.error(f"Erro ao editar mensagem no Telegram: {e}")
 
 def send_telegram_file(file_path, caption=""):
     try:
@@ -49,33 +54,46 @@ def send_telegram_file(file_path, caption=""):
             files = {"document": file}
             requests.post(url, data=data, files=files)
     except Exception as e:
-        print(f"Erro ao enviar arquivo ao Telegram: {e}")
+        logging.error(f"Erro ao enviar arquivo ao Telegram: {e}")
 
-# Função para capturar informações do lunch
-def get_lunch_info():
-    try:
-        process = subprocess.run(
-            f"source build/envsetup.sh && lunch {LUNCH_TARGET}",
+# Pegar infromações sobre a ROM e device
+def get_rom_info():
+    process = subprocess.run(
+        f"source build/envsetup.sh && lunch {LUNCH_TARGET}",
+        shell=True,
+        executable="/bin/bash",
+        capture_output=True,
+        text=True,
+    )
+    output = process.stdout
+    rom_info = re.search(r"CUSTOM_VERSION=([\w\-.]+)", output)
+    android_version = re.search(r"PLATFORM_VERSION=([\d.]+)", output)
+    device = re.search(r"TARGET_PRODUCT=(\w+)", output)
+
+    rom = rom_info.group(1) if rom_info else "Desconhecido"
+    version = android_version.group(1) if android_version else "Desconhecido"
+    device_name = device.group(1).split("_")[1] if device else "Desconhecido"
+
+    return rom, version, device_name
+
+# Função para iniciar o processo de build
+def start_build():
+    build_command = f"source build/envsetup.sh && lunch {LUNCH_TARGET} && make {BUILD_TARGET} -j$(nproc)"
+    log_file = "build.log"
+
+    with open(log_file, "w") as log:
+        build_process = subprocess.Popen(
+            build_command,
             shell=True,
             executable="/bin/bash",
-            capture_output=True,
-            text=True,
+            stdout=log,
+            stderr=log
         )
-        output = process.stdout
-        rom_info = re.search(r"CUSTOM_VERSION=([\w\-.]+)", output)
-        android_version = re.search(r"PLATFORM_VERSION=([\d.]+)", output)
-        device = re.search(r"TARGET_PRODUCT=(\w+)", output)
 
-        rom = rom_info.group(1) if rom_info else "Desconhecido"
-        version = android_version.group(1) if android_version else "Desconhecido"
-        device_name = device.group(1).split("_")[1] if device else "Desconhecido"
+    return build_process, log_file
 
-        return rom, version, device_name
-    except Exception as e:
-        print(f"Erro ao obter informações do lunch: {e}")
-        return "Desconhecido", "Desconhecido", "Desconhecido"
-
-def monitor_system_resources():
+# Capturar uso de recursos do sistema
+def get_system_resources():
     try:
         cpu_usage = psutil.cpu_percent(interval=1)
         memory = psutil.virtual_memory()
@@ -86,11 +104,10 @@ def monitor_system_resources():
             "disk": f"{disk.used // (1024 ** 3)}GB/{disk.total // (1024 ** 3)}GB ({disk.percent}%)"
         }
     except Exception as e:
-        print(f"Erro ao monitorar recursos do sistema: {e}")
+        logging.error(f"Erro ao monitorar recursos do sistema: {e}")
         return {"cpu": "N/A", "ram": "N/A", "disk": "N/A"}
 
-
-# Função para monitorar progresso do build
+# Monitorar progresso do build e enviar para o telegram
 def monitor_build_progress(log_file, message_id, rom, version, device_name):
     progress_pattern = re.compile(r"\[\s*(\d+)%\s+(\d+/\d+)\s+([\d\w]+ remaining)\]")  # Captura progresso detalhado
     previous_progress = None
@@ -109,7 +126,7 @@ def monitor_build_progress(log_file, message_id, rom, version, device_name):
                     progress = f"{percentage} {tasks} {time_remaining}"
                     if progress != previous_progress:
                         # Monitorar recursos do sistema
-                        resources = monitor_system_resources()
+                        resources = get_system_resources()
                         cpu = resources["cpu"]
                         ram = resources["ram"]
                         disk = resources["disk"]
@@ -131,8 +148,9 @@ def monitor_build_progress(log_file, message_id, rom, version, device_name):
                 if "ota_from_target_files.py - INFO    : done" in line:
                     break
     except Exception as e:
-        print(f"Erro ao monitorar progresso: {e}")
+        logging.error(f"Erro ao monitorar progresso: {e}")
 
+# Pixeldrain API
 def upload_file_to_pixeldrain(rom_path, API_KEY):
     url = "https://pixeldrain.com/api/file/"
 
@@ -142,28 +160,29 @@ def upload_file_to_pixeldrain(rom_path, API_KEY):
             capture_output=True,
             text=True
         )
-        
+
         if result.returncode != 0:
-            print(f"Erro ao fazer upload para o Pixeldrain: {result.stderr}")
+            logging.error(f"Erro ao fazer upload para o Pixeldrain: {result.stderr}")
             return None
-        
+
         response_data = json.loads(result.stdout)
-        print(f"Resposta da API do Pixeldrain: {response_data}")  # Adicionar este print para depuração
+        logging.info(f"Resposta da API do Pixeldrain: {response_data}")  # Adicionar este print para depuração
         file_id = response_data.get("id")
 
         if file_id:
             return f"https://pixeldrain.com/u/{file_id}"
         else:
-            print("Erro: Não foi possível obter o ID do arquivo do Pixeldrain.")
+            logging.error("Erro: Não foi possível obter o ID do arquivo do Pixeldrain.")
             return None
 
     except FileNotFoundError:
-        print(f"Erro: O arquivo '{rom_path}' não foi encontrado.")
+        logging.error(f"Erro: O arquivo '{rom_path}' não foi encontrado.")
         return None
     except Exception as e:
-        print(f"Erro ao fazer upload para o Pixeldrain: {e}")
+        logging.error(f"Erro ao fazer upload para o Pixeldrain: {e}")
         return None
 
+# Fazer upload da ROM compilada
 def upload_build(device_name, rom, log_file):
     # Obter path do arquivo
     rom_path = None
@@ -185,7 +204,7 @@ def upload_build(device_name, rom, log_file):
 
     # Verificar se o arquivo ROM existe
     if not os.path.exists(rom_path):
-        print(f"Erro: O arquivo ROM '{rom_path}' não foi encontrado.")
+        logging.error(f"Erro: O arquivo ROM '{rom_path}' não foi encontrado.")
         edit_telegram_message(
             upload_message_id,
             f"🔴 <b>Falha no upload: arquivo ROM não encontrado!</b>\n"
@@ -210,7 +229,7 @@ def upload_build(device_name, rom, log_file):
         )
         return True
     else:
-        print("Erro: Falha ao fazer upload do arquivo.")
+        logging.error("Erro: Falha ao fazer upload do arquivo.")
         edit_telegram_message(
             upload_message_id,
             f"🔴 <b>Falha no upload: falha ao fazer upload do arquivo!</b>\n"
@@ -222,10 +241,10 @@ def upload_build(device_name, rom, log_file):
 
 # Função principal
 def main():
-    log_file = "build.log"
+    log_file = None  # Inicialize com None para evitar erros de referência
     try:
-        # Capturar informações do lunch
-        rom, version, device_name = get_lunch_info()
+        # Iniciar lunch e obter informações sobre a ROM/device
+        rom, version, device_name = get_rom_info()
 
         # Enviar mensagem inicial e obter o ID da mensagem
         message_id = send_telegram_message(
@@ -236,28 +255,18 @@ def main():
             f"<b>Progresso:</b> 0%"
         )
 
-        # Criar o comando para configurar o ambiente e executar o build
-        build_command = f"""
-        source build/envsetup.sh && \
-        lunch {LUNCH_TARGET} && \
-        make {BUILD_TARGET} -j$(nproc)
-        """
+        # Iniciar o build e monitorar o progresso em paralelo
+        build_process, log_file = start_build()
 
-        # Iniciar o build e monitorar o progresso
-        with open(log_file, "w") as log:
-            build_process = subprocess.Popen(
-                build_command,
-                shell=True,
-                executable="/bin/bash",
-                stdout=log,
-                stderr=log
-            )
+        # Criar uma thread para monitorar o progresso
+        monitor_thread = threading.Thread(
+            target=monitor_build_progress,
+            args=(log_file, message_id, rom, version, device_name)
+        )
+        monitor_thread.start()
 
-            # Monitorar progresso em paralelo
-            monitor_build_progress(log_file, message_id, rom, version, device_name)
-
-            # Aguardar conclusão do build
-            build_process.wait()
+        # Aguardar conclusão do build
+        build_process.wait()
 
         # Verificar resultado do build
         if build_process.returncode == 0:
@@ -268,16 +277,20 @@ def main():
                 f"<b>Android:</b> {version}\n"
                 f"<b>Dispositivo:</b> {device_name}",
             )
-            time.sleep(5) # Aguardar 5 segundos antes de fazer upload
+            time.sleep(5)  # Aguardar 5 segundos antes de fazer upload
             upload_build(device_name, rom, log_file)
         else:
             edit_telegram_message(message_id, "🔴 <b>Compilação falhou!</b>")
             send_telegram_file(log_file, "🔴 <b>Log de erro:</b>")
 
+        # Aguarde o término da thread de monitoramento (opcional)
+        monitor_thread.join()
+
     except Exception as e:
         send_telegram_message(f"🔴 <b>Erro inesperado:</b> {str(e)}")
     finally:
-        if os.path.exists(log_file):
+        # Verifique se log_file foi definido e existe antes de usá-lo
+        if log_file and os.path.exists(log_file):
             send_telegram_file(log_file, "📄 <b>Log final:</b>")
             os.remove(log_file)
